refactor(dbt): migrate models from SQLite to Postgres syntax
Migrate the dbt project to run against Supabase Postgres instead of
the local SQLite database:

- profiles.yml: switch the dev target to type=postgres, reading
  connection params from SUPABASE_* env vars with sslmode=require.
- models/staging/stg_sales.sql: cast price to NUMERIC instead of
  FLOAT. Postgres has no ROUND(double precision, integer), which
  was breaking the downstream revenue calculation in int_sales_revenue.
- models/staging/schema.yml: declare schema=clearpath on the source
  so dbt looks in the correct Postgres schema (the SQLite default
  of "main" doesn't apply).
- models/intermediate/int_sales_revenue.sql: rename `revenue` to
  `total_revenue` so the column name matches what schema.yml
  asserts (the schema test was previously masked by SQLite's
  permissive name resolution).
- models/marts/mart_daily_revenue.sql: replace SQLite's
  date(..., '-30 days') modifier with `... - INTERVAL '30 days'`.
- models/marts/mart_product_velocity.sql: replace julianday()
  arithmetic with native date subtraction (Postgres returns an
  integer number of days when subtracting two date values).
- models/marts/mart_top_products.sql: drop the `date` column from
  the SELECT. It wasn't in GROUP BY (Postgres rejects this; SQLite
  silently allowed it) and was never part of the model's declared
  schema.

After this commit: `python -m dbt.cli.main run` passes 5 of 5 models
and `python -m dbt.cli.main test` passes 22 of 22 tests against
Supabase.

+++ clearpath_dbt/models/staging/stg_sales.sql
@@ -9,7 +9,7 @@ SELECT
     category,
     size,
     CAST(quantity AS INTEGER) as quantity,
-    CAST(price AS FLOAT) as price,
+    CAST(price AS NUMERIC) as price,
     is_known_product
 FROM {{ source('main', 'sales') }}
 WHERE quantity > 0
